requirements.txt
django
django-phonenumber-field[phonenumbers]
Babel
hashids